Fix TomlStringConfigEntry.java reading null-bytes
The previous implementation of the read(ByteBuffer buffer) method was initializing a new byte array without actually reading data from the ByteBuffer. This resulted in a string of null characters. 

This commit fixes the issue by adding a buffer.get(bytes) call to read the bytes from the ByteBuffer into the byte array before converting it to a string using UTF-8 encoding.